slib: Add form_fill handling

+++ slib/attach.go
@@ -11,12 +11,17 @@ package slib
 import (
    "fmt"
    "io"
+   "encoding/json"
    "net"
    "os"
    "path"
    "strings"
    "syscall"
 )
+
+const kSuffixRecv = "_recv"
+const kSuffixSent = "_sent"
+
 
 type tAttachEl struct {
    Name, MsgId, File string
@@ -56,35 +61,57 @@ func GetPathAttach(iSvc string, iState *ClientState, iFile string) string {
 
 func attachSub(iSvc, iSub string) string { return attachDir(iSvc) + iSub + "/" }
 
+func makeAttach(i *Update) []tHeader2Attach {
+   aAtc := make([]tHeader2Attach, len(i.Thread.Attach))
+   for a, aName := range i.Thread.Attach {
+      aAtc[a].Name = aName
+      if strings.HasPrefix(aName, "form_fill/") {
+         aAtc[a].Size = int64(len(i.Thread.FormFill[aName[10:]]))
+      }
+   }
+   return aAtc
+}
+
 func sizeSavedAttach(iSvc string, iSubHead *tHeader2, iId tSaveId) int64 {
    aTid := iId.tid(); if aTid == "" { aTid = "_" + iId.sid() }
    aPrefix := attachSub(iSvc, aTid) + iId.sid() + "_"
    var aTotal int64
 
    for a, aFile := range iSubHead.Attach {
-      aFi, err := os.Lstat(aPrefix + path.Base(aFile.Name))
-      if err != nil { quit(err) }
-      iSubHead.Attach[a].Size = aFi.Size()
-      iSubHead.Attach[a].Name = path.Base(aFile.Name)
-      aTotal += aFi.Size()
+      if aFile.Size == 0 {
+         aFi, err := os.Lstat(aPrefix + path.Base(aFile.Name))
+         if err != nil { quit(err) }
+         iSubHead.Attach[a].Size = aFi.Size()
+         iSubHead.Attach[a].Name = path.Base(aFile.Name)
+      }
+      aTotal += iSubHead.Attach[a].Size
    }
    return aTotal
 }
 
-func sendSavedAttach(iConn net.Conn, iSvc string, iSubHead *tHeader2, iId tSaveId) error {
+func sendSavedAttach(iConn net.Conn, iSvc string, iSubHead *tHeader2, iId tSaveId, iFd *os.File) error {
+   var err error
    aTid := iId.tid(); if aTid == "" { aTid = "_" + iId.sid() }
    aPrefix := attachSub(iSvc, aTid) + iId.sid() + "_"
    for _, aFile := range iSubHead.Attach {
-      aFd, err := os.Open(aPrefix + path.Base(aFile.Name))
-      if err != nil { quit(err) }
-      _, err = io.Copy(iConn, aFd)
-      aFd.Close()
+      var aXd, aFd *os.File = iFd, nil
+      var aFi os.FileInfo
+      if !strings.HasPrefix(aFile.Name, "form_fill/") {
+         aFd, err = os.Open(aPrefix + path.Base(aFile.Name))
+         if err != nil { quit(err) }
+         defer aFd.Close()
+         aFi, err = aFd.Stat()
+         if err != nil { quit(err) }
+         if aFi.Size() != aFile.Size { quit(tError("file size mismatch")) }
+         aXd = aFd
+      }
+      _, err = io.CopyN(iConn, aXd, aFile.Size)
       if err != nil { return err } //todo only return net errors
    }
    return nil
 }
 
-func writeReceivedAttach(iSvc string, iHead *Header, iData []byte, iR io.Reader) error {
+func tempReceivedAttach(iSvc string, iHead *Header, iData []byte, iR io.Reader) error {
    var err error
    aWritten := iHead.DataLen - totalAttach(&iHead.SubHead)
    if aWritten >= int64(len(iData)) {
@@ -92,10 +119,19 @@ func writeReceivedAttach(iSvc string, iHead *Header, iData []byte, iR io.Reader)
    } else {
       iData = iData[aWritten:]
    }
+   aDoSync := false
    for _, aFile := range iHead.SubHead.Attach {
-      aPath := tempDir(iSvc) + iHead.Id + "_" + aFile.Name //todo escape '/' in .Name
+      aDoSync = true
+      if strings.HasPrefix(aFile.Name, "form_fill/") {
+         aTid := iHead.SubHead.ThreadId; if aTid == "" { aTid = iHead.Id }
+         err = tempForm(iSvc, aTid, iHead.Id, kSuffixRecv, &aFile, iData, iR)
+         if err != nil { return err }
+         if aFile.Size >= int64(len(iData)) { iData = nil } else { iData = iData[aFile.Size:] }
+         continue
+      }
+      aPath := tempDir(iSvc) + iHead.Id + "_" + aFile.Name + ".tmp" //todo escape '/' in .Name
       var aFd *os.File
-      aFd, err = os.OpenFile(aPath + ".tmp", os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0600)
+      aFd, err = os.OpenFile(aPath, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0600)
       if err != nil { quit(err) }
       defer aFd.Close()
       var aLen int64 = 0
@@ -108,19 +144,17 @@ func writeReceivedAttach(iSvc string, iHead *Header, iData []byte, iR io.Reader)
       if aLen < aFile.Size {
          _, err = io.CopyN(aFd, iR, aFile.Size - aLen)
          if err != nil {
-            os.Remove(aPath + ".tmp")
+            os.Remove(aPath)
             return err //todo only return net errors
          }
       }
       err = aFd.Sync()
       if err != nil { quit(err) }
-      err = os.Remove(aPath + ".atc")
-      if err != nil && !os.IsNotExist(err) { quit(err) }
-      err = os.Rename(aPath + ".tmp", aPath + ".atc")
+   }
+   if aDoSync {
+      err = syncDir(tempDir(iSvc))
       if err != nil { quit(err) }
    }
-   err = syncDir(tempDir(iSvc))
-   if err != nil { quit(err) }
    return nil
 }
 
@@ -136,13 +170,34 @@ func storeReceivedAttach(iSvc string, iSubHead *tHeader2, iRec tComplete) {
       err = syncDir(attachDir(iSvc))
       if err != nil { quit(err) }
    }
+   aDoSync := false
    for _, aFile := range iSubHead.Attach {
-      err = renameRemove(tempDir(iSvc) + iRec.mid() + "_" + aFile.Name + ".atc",
+      if strings.HasPrefix(aFile.Name, "form_fill/") { continue }
+      aDoSync = true
+      err = renameRemove(tempDir(iSvc) + iRec.mid() + "_" + aFile.Name + ".tmp",
                          attachSub(iSvc, iRec.tid()) + iRec.mid() + "_" + aFile.Name)
       if err != nil { quit(err) }
    }
-   err = syncDir(attachSub(iSvc, iRec.tid()))
-   if err != nil { quit(err) }
+   if aDoSync {
+      err = syncDir(attachSub(iSvc, iRec.tid()))
+      if err != nil { quit(err) }
+   }
+   _storeFormAttach(iSvc, iSubHead, iRec)
+}
+
+func tempSavedAttach(iSvc string, iHead *Header, iSd *os.File) {
+   var err error
+   aDoSync := false
+   for _, aFile := range iHead.SubHead.Attach {
+      if !strings.HasPrefix(aFile.Name, "form_fill/") { continue }
+      aDoSync = true
+      err = tempForm(iSvc, iHead.SubHead.ThreadId, iHead.Id, kSuffixSent, &aFile, nil, iSd)
+      if err != nil { quit(err) }
+   }
+   if aDoSync {
+      err = syncDir(tempDir(iSvc))
+      if err != nil { quit(err) }
+   }
 }
 
 func storeSavedAttach(iSvc string, iSubHead *tHeader2, iRec tComplete) {
@@ -156,24 +211,74 @@ func storeSavedAttach(iSvc string, iSubHead *tHeader2, iRec tComplete) {
       err = syncDir(attachDir(iSvc))
       if err != nil { quit(err) }
    }
+   aDoSync := false
    for _, aFile := range iSubHead.Attach {
+      if strings.HasPrefix(aFile.Name, "form_fill/") { continue }
+      aDoSync = true
       err = renameRemove(attachSub(iSvc, iRec.tid()) + iRec.sid() + "_" + path.Base(aFile.Name),
                          attachSub(iSvc, iRec.tid()) + iRec.mid() + "_" + path.Base(aFile.Name))
       if err != nil { quit(err) }
    }
-   err = syncDir(attachSub(iSvc, iRec.tid()))
-   if err != nil { quit(err) }
+   if aDoSync {
+      err = syncDir(attachSub(iSvc, iRec.tid()))
+      if err != nil { quit(err) }
+   }
+   _storeFormAttach(iSvc, iSubHead, iRec)
+}
+
+func _storeFormAttach(iSvc string, iSubHead *tHeader2, iRec tComplete) {
+   aSuffix := kSuffixRecv; if iRec.sid() != "" { aSuffix = kSuffixSent }
+   aDoSync := false
+   for _, aFile := range iSubHead.Attach {
+      if !strings.HasPrefix(aFile.Name, "form_fill/") { continue }
+      aOk := storeForm(iSvc, iRec.mid(), aSuffix, &aFile)
+      aDoSync = aDoSync || aOk
+   }
+   if aDoSync {
+      err := syncDir(formDir(iSvc))
+      if err != nil { quit(err) }
+   }
 }
 
 func validateSavedAttach(iSvc string, iSubHead *tHeader2, iId tSaveId) error {
    aTid := iId.tid(); if aTid == "" { aTid = "_" + iId.sid() }
    for _, aFile := range iSubHead.Attach {
+      if strings.HasPrefix(aFile.Name, "form_fill/") { continue }
       _, err := os.Lstat(attachSub(iSvc, aTid) + iId.sid() + "_" + path.Base(aFile.Name))
       if err != nil {
          return tError(fmt.Sprintf("%s missing %s", aTid, aFile.Name))
       }
    }
    return nil
+}
+
+func writeFormFillAttach(iFd *os.File, iSubHead *tHeader2, iMap map[string]string, iEl *tIndexEl) {
+   var err error
+   aCw := tCrcWriter{sum:iEl.Checksum}
+   aTee := io.MultiWriter(iFd, &aCw)
+   aSize := iEl.Size
+
+   for _, aFile := range iSubHead.Attach {
+      if !strings.HasPrefix(aFile.Name, "form_fill/") { continue }
+      if iEl.Size == aSize {
+         _, err = iFd.Seek(-1, io.SeekCurrent) // overwrite '\n'
+         if err != nil { quit(err) }
+      }
+      aS := []byte(iMap[aFile.Name[10:]])
+      if int64(len(aS)) != aFile.Size || aFile.Size <= 0 {
+         quit(tError("empty or mis-sized " + aFile.Name))
+      }
+      err = json.Unmarshal(aS, &struct{}{})
+      if err != nil { quit(err) }
+      _, err = aTee.Write(aS)
+      if err != nil { quit(err) }
+      iEl.Size += aFile.Size
+      iEl.Checksum = aCw.sum
+   }
+   if iEl.Size > aSize {
+      _, err = iFd.Write([]byte{'\n'})
+      if err != nil { quit(err) }
+   }
 }
 
 func updateSavedAttach(iSvc string, iSubHeadOld, iSubHeadNew *tHeader2, iRec tComplete) {
@@ -184,6 +289,7 @@ func updateSavedAttach(iSvc string, iSubHeadOld, iSubHeadNew *tHeader2, iRec tCo
 
    if aHasOld {
       for _, aFile := range iSubHeadOld.Attach {
+         if strings.HasPrefix(aFile.Name, "form_fill/") { continue }
          err = os.Remove(attachSub(iSvc, aTid) + iRec.sid() + "_" + path.Base(aFile.Name))
          if err != nil && !os.IsNotExist(err) { quit(err) }
       }
@@ -207,6 +313,7 @@ func updateSavedAttach(iSvc string, iSubHeadOld, iSubHeadNew *tHeader2, iRec tCo
          if err != nil { quit(err) }
       }
       for _, aFile := range iSubHeadNew.Attach {
+         if strings.HasPrefix(aFile.Name, "form_fill/") { continue }
          err = os.Link(kStorageDir + aFile.Name,
                        attachSub(iSvc, aTid) + iRec.sid() + "_" + path.Base(aFile.Name))
          if err != nil {
@@ -222,6 +329,7 @@ func updateSavedAttach(iSvc string, iSubHeadOld, iSubHeadNew *tHeader2, iRec tCo
 }
 
 func totalAttach(iSubHead *tHeader2) int64 {
+   if iSubHead.isSaved { return 0 }
    var aLen int64
    for _, aFile := range iSubHead.Attach { aLen += aFile.Size }
    return aLen
